--- CommcareHQ/TestScripts/cleanup.py
@@ -1,47 +1,0 @@
-import unittest
-from CommcareHQ.Pages.groupPage import GroupPage
-from CommcareHQ.Pages.mobileWorkersPage import MobileWorkerPage
-from CommcareHQ.Pages.rolesPermissionsPage import RolesPermissionPage
-from CommcareHQ.TestBase.environmentSetupPage import EnvironmentSetup
-
-
-class CleanUps(EnvironmentSetup):
-
-    def test_01_cleanup_mobile_worker(self):
-        driver = self.driver
-        clean = MobileWorkerPage(driver)
-        clean.mobile_worker_menu()
-        clean.select_mobile_worker_created()
-        clean.cleanup_mobile_worker()
-        print("Deleted the mobile worker")
-
-    def test_02_cleanup_user_field(self):
-        driver = self.driver
-        clean = MobileWorkerPage(driver)
-        clean.mobile_worker_menu()
-        clean.edit_user_field()
-        clean.cleanup_user_field()
-        clean.save_field()
-        print("Deleted the user field")
-
-    def test_03_cleanup_group(self):
-        driver = self.driver
-        clean = GroupPage(driver)
-        clean2 = MobileWorkerPage(driver)
-        clean2.mobile_worker_menu()
-        clean.click_group_menu()
-        clean.cleanup_group()
-        print("Deleted the group")
-
-    def test_04_cleanup_role(self):
-        driver = self.driver
-        clean = RolesPermissionPage(driver)
-        clean2 = MobileWorkerPage(driver)
-        clean2.mobile_worker_menu()
-        clean.roles_menu_click()
-        clean.cleanup_role()
-        print("Deleted the role")
-
-
-if __name__ == "__main__":
-    unittest.main()
