+++ UserInputs/loginCredentials.py
@@ -1,0 +1,5 @@
+class LoginCredentials:
+    # inputs for login page
+    url = ""
+    login_username = ""
+    login_password = ""
